Moved AI implementation to Study Mode Section
