+++ src/main/java/QUT/CAB302/fortunecookie/HomepageController.java
@@ -1,6 +1,5 @@
 package QUT.CAB302.fortunecookie;
 
-import javafx.application.Platform;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
@@ -14,14 +13,6 @@ import javafx.animation.ScaleTransition;
 import javafx.animation.FadeTransition;
 import javafx.util.Duration;
 import javafx.scene.Node;
-
-import org.json.JSONObject;
-import javafx.scene.control.TextArea;
-import java.io.BufferedReader;
-import java.io.InputStreamReader;
-import java.net.HttpURLConnection;
-import java.net.URL;
-import java.io.OutputStream;
 
 public class HomepageController {
 
@@ -165,57 +156,5 @@ public class HomepageController {
             alert.setContentText("Could not load the Study Vault page. Please try again.");
             alert.showAndWait();
         }
-    }
-
-    // AI Implementation
-
-    @FXML
-    private TextArea aiInput;
-    @FXML
-    private TextArea aiResponse;
-
-    @FXML
-    public void handleAskAI() {
-        String model = "mistral";
-        String prompt = aiInput.getText();
-        String line;
-        Runnable task = new Runnable() {
-            @Override
-            public void run() {
-
-                try {
-                    // Set up an HTTP POST request
-                    URL url = new URL("http://localhost:11434/api/generate");
-                    HttpURLConnection conn = (HttpURLConnection) url.openConnection();
-                    conn.setRequestMethod("POST");
-                    conn.setRequestProperty("Content-Type", "application/json");
-                    conn.setDoOutput(true);
-
-                    // Create request JSON
-                    JSONObject requestJson = new JSONObject();
-                    requestJson.put("model", model);
-                    requestJson.put("prompt", prompt);
-                    requestJson.put("stream", false);
-
-                    // Send request
-                    try (OutputStream os = conn.getOutputStream()) {
-                        os.write(requestJson.toString().getBytes());
-                    }
-
-                    // Get response
-                    try (BufferedReader br = new BufferedReader(new InputStreamReader(conn.getInputStream()))) {
-                        String responseLine = br.readLine();
-                        JSONObject responseJson = new JSONObject(responseLine);
-                        String fullResponse = responseJson.getString("response");
-
-                        Platform.runLater(() -> aiResponse.setText(fullResponse));
-                    }
-                } catch (Exception e) {
-                    e.printStackTrace();
-                    Platform.runLater(() -> aiResponse.setText("Error: " + e.getMessage()));
-                }
-            }
-        };
-        new Thread(task).start();
     }
 }
