+++ src/main/resources/QUT/CAB302/fortunecookie/studymode.fxml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<?import javafx.geometry.Pos?>
+
 <?import javafx.scene.control.Button?>
 <?import javafx.scene.control.Label?>
 <?import javafx.scene.layout.BorderPane?>
@@ -10,9 +10,8 @@
 <?import javafx.scene.control.TextField?>
 <?import javafx.scene.control.ComboBox?>
 <?import javafx.scene.layout.Region?>
-<?import javafx.collections.FXCollections?>
-<?import java.lang.String?>
 
+<?import javafx.scene.control.TextArea?>
 <BorderPane xmlns="http://javafx.com/javafx/23.0.1"
             xmlns:fx="http://javafx.com/fxml/1"
             fx:controller="QUT.CAB302.fortunecookie.StudyModeController"
@@ -53,6 +52,9 @@
                     <Label text="Select Your Mood:" style="-fx-font-size: 14px;"/>
                     <ComboBox fx:id="moodComboBox" promptText="Choose mood" style="-fx-font-size: 14px;"/>
                 </HBox>
+
+                <Button text="Ask AI" onAction="#handleAskAI"/>
+                <TextArea fx:id="aiResponse" editable="false" wrapText="true" prefRowCount="5"/>
 
                 <!-- Timer Display -->
                 <Label text="00:00" fx:id="timerLabel"
